Add command palette item for building all
