+++ src/main/kotlin/org/elm/ide/toolwindow/ElmCompilerToolWindowFactory.kt
@@ -31,7 +31,6 @@ import com.intellij.util.ui.UIUtil
 import org.elm.ide.notifications.showBalloon
 import org.elm.workspace.ElmProject
 import org.elm.workspace.ElmWorkspaceService
-import org.elm.workspace.commandLineTools.makeAllTargets
 import org.elm.workspace.commandLineTools.makeProject
 import org.elm.workspace.compiler.*
 import org.elm.workspace.elmWorkspace
@@ -71,7 +70,7 @@ class ElmCompilerToolWindowFactory : ToolWindowFactory {
             isBuildTargetsVisible = { buildTargetsVisible },
             onBuildSelected = { buildTargetsPanelRef.buildSelectedTarget() },
             isBuildSelectedEnabled = { buildTargetsPanelRef.hasSelectedTarget() },
-            onBuildAll = { buildTargetsPanelRef.buildAllTargets() },
+            onBuildAll = { org.elm.ide.actions.buildAllTargets(project) },
             isBuildAllEnabled = { buildTargetsPanelRef.hasTargets() }
         )
         val outputPanel = ElmCompilerOutputPanel(project)
@@ -204,36 +203,6 @@ private class ElmBuildTargetsPanel(private val project: Project) : JPanel(Border
     fun hasSelectedTarget(): Boolean = targetList.selectedIndex >= 0
 
     fun hasTargets(): Boolean = !targetListModel.isEmpty
-
-    fun buildAllTargets() {
-        val currentFileInEditor: VirtualFile? = FileEditorManager.getInstance(project).selectedFiles.firstOrNull()
-        val targetsByProject = mutableListOf<Pair<ElmProject, List<ResolvedBuildTarget>>>()
-        for (elmProject in project.elmWorkspace.allProjects.sortedBy { it.presentableName }) {
-            val resolved = when (val result = project.elmWorkspace.resolveBuildTargets(elmProject)) {
-                is org.elm.openapiext.Result.Ok -> result.value
-                is org.elm.openapiext.Result.Err -> emptyList()
-            }
-            if (resolved.isNotEmpty()) targetsByProject += elmProject to resolved
-        }
-        if (targetsByProject.isEmpty()) return
-
-        ApplicationManager.getApplication().executeOnPooledThread {
-            val existingByProject = targetsByProject.mapNotNull { (elmProject, targets) ->
-                val existing = targets.filter { Files.exists(it.inputPath) }
-                if (existing.isEmpty()) null else elmProject to existing
-            }
-            if (existingByProject.isEmpty()) {
-                ApplicationManager.getApplication().invokeLater {
-                    project.showBalloon(
-                        "Cannot build targets: no build target input files were found.",
-                        NotificationType.ERROR
-                    )
-                }
-                return@executeOnPooledThread
-            }
-            makeAllTargets(project, existingByProject, currentFileInEditor)
-        }
-    }
 
     fun buildSelectedTarget() {
         val item = targetList.selectedValue ?: return
